requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.5.0
Django==4.1.1
django-extensions==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
facebook-sdk==3.1.0
future==0.18.2
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
oauthlib==3.2.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.5.0
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.21.0
python-twitter==3.5
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
